move route_agent to its own directory module
