+++ examples/guide/src/guide.rs
@@ -5,7 +5,7 @@ use yew::prelude::*;
 use yew::Properties;
 use yew_router::components::RouterLink;
 use yew_router::matcher::RouteMatcher;
-use yew_router::route_agent::RouteRequest::GetCurrentRoute;
+use yew_router::agent::RouteRequest::GetCurrentRoute;
 use yew_router::{RouteAgent, RouteInfo};
 
 pub struct Guide {
